chore: move exception to utils, better message
